Add unit more unit-tests
